REFACTOR test_conduit suites to single test module
